+++ internal/provider/provider.go
@@ -153,7 +153,7 @@ func (p *GroundcoverProvider) Resources(ctx context.Context) []func() resource.R
 		NewMonitorResource,
 		NewApiKeyResource,
 		NewLogsPipelineResource,
-		NewMetricsAggregatorResource,
+		NewMetricsAggregationResource,
 		NewIngestionKeyResource,
 		NewDashboardResource,
 		NewDataIntegrationResource,